+++ plugins/src/main/java/org/smartregister/fhir/gateway/plugins/SyncAccessDecision.java
@@ -20,7 +20,6 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.http.HttpResponse;
 import org.apache.http.impl.client.BasicResponseHandler;
 import org.apache.http.util.TextUtils;
-import org.hl7.fhir.instance.model.api.IBaseBundle;
 import org.hl7.fhir.instance.model.api.IBaseResource;
 import org.hl7.fhir.r4.model.Bundle;
 import org.hl7.fhir.r4.model.ListResource;
@@ -360,49 +359,8 @@ public class SyncAccessDecision implements AccessDecision {
         StringBuilder urlBuilder = new StringBuilder(request.getFhirServerBase());
         urlBuilder.append("/").append(request.getRequestPath());
 
-        return addPaginationLinks(urlBuilder, resultBundle, page, totalEntries, count, parameters);
-    }
-
-    public static Bundle addPaginationLinks(
-            StringBuilder urlBuilder,
-            Bundle resultBundle,
-            int page,
-            int totalEntries,
-            int count,
-            Map<String, String[]> parameters) {
-        resultBundle.setTotal(totalEntries);
-
-        int nextPage =
-                page < ((float) totalEntries / count) ? page + 1 : 0; // 0 indicates no next page
-        int prevPage = page > 1 ? page - 1 : 0; // 0 indicates no previous page
-
-        Bundle.BundleLinkComponent selfLink = new Bundle.BundleLinkComponent();
-        List<Bundle.BundleLinkComponent> link = new ArrayList<>();
-        String selfUrl = constructUpdatedUrl(new StringBuilder(urlBuilder), parameters);
-        selfLink.setRelation(IBaseBundle.LINK_SELF);
-        selfLink.setUrl(selfUrl);
-        link.add(selfLink);
-        resultBundle.setLink(link);
-
-        if (nextPage > 0) {
-            parameters.put(
-                    Constants.PAGINATION_PAGE_NUMBER, new String[] {String.valueOf(nextPage)});
-            String nextUrl = constructUpdatedUrl(new StringBuilder(urlBuilder), parameters);
-            Bundle.BundleLinkComponent nextLink = new Bundle.BundleLinkComponent();
-            nextLink.setRelation(IBaseBundle.LINK_NEXT);
-            nextLink.setUrl(nextUrl);
-            resultBundle.addLink(nextLink);
-        }
-        if (prevPage > 0) {
-            parameters.put(
-                    Constants.PAGINATION_PAGE_NUMBER, new String[] {String.valueOf(prevPage)});
-            String prevUrl = constructUpdatedUrl(new StringBuilder(urlBuilder), parameters);
-            Bundle.BundleLinkComponent previousLink = new Bundle.BundleLinkComponent();
-            previousLink.setRelation(IBaseBundle.LINK_PREV);
-            previousLink.setUrl(prevUrl);
-            resultBundle.addLink(previousLink);
-        }
-        return resultBundle;
+        return Utils.addPaginationLinks(
+                urlBuilder, resultBundle, page, totalEntries, count, parameters);
     }
 
     private String getSyncTagUrl(String syncStrategy) {
@@ -483,26 +441,6 @@ public class SyncAccessDecision implements AccessDecision {
         }
 
         return false;
-    }
-
-    private static String constructUpdatedUrl(
-            StringBuilder urlBuilder, Map<String, String[]> parameters) {
-        urlBuilder.append("?");
-        for (Map.Entry<String, String[]> entry : parameters.entrySet()) {
-            String paramName = entry.getKey();
-            String[] paramValues = entry.getValue();
-
-            for (String paramValue : paramValues) {
-                urlBuilder.append(paramName).append("=").append(paramValue).append("&");
-            }
-        }
-
-        // Remove the trailing '&' if present
-        if (urlBuilder.charAt(urlBuilder.length() - 1) == '&') {
-            urlBuilder.deleteCharAt(urlBuilder.length() - 1);
-        }
-
-        return urlBuilder.toString();
     }
 
     private boolean isResourceTypeRequest(String requestPath) {
